+++ src/module/sdk/NearApiService/NearBlocks/NearBlocksService.ts
@@ -17,11 +17,10 @@ import { FetchService } from "../common/FetchService";
 import { NearApiServiceInterface, NearApiServicePaginatedParams, NearApiServiceParams } from "../NearApiService.types";
 import {
     NearBlocksTransactionResponseDto,
-    NearblocksAccessKeyResponseDto,
-    NearBlocksTokenResponseDto,
     NearBlocksTransactionDto,
     NearBlocksActionDto,
     NearBlocksKitWalletStakingDepositsResponseDto,
+    NearBlocksActivityDto,
 } from "./NearBlocksService.types";
 
 export class NearBlocksService extends FetchService implements NearApiServiceInterface {
@@ -112,33 +111,55 @@ export class NearBlocksService extends FetchService implements NearApiServiceInt
     }
 
     async getAccountsFromPublicKey({ address }: NearApiServiceParams): Promise<string[]> {
-        const accounts: string[] = [];
-        const keys = await this.fetch<NearblocksAccessKeyResponseDto>(`/keys/${address}`);
-        if (!keys?.keys || keys.keys.length === 0) return accounts;
-        for (const key of keys.keys) {
-            if (key.permission_kind === "FULL_ACCESS") {
-                accounts.push(key.account_id);
-            }
-        }
-        return accounts;
-    }
-
-    private async getAccountTokens({ address }: NearApiServiceParams): Promise<NearBlocksTokenResponseDto> {
-        return await this.fetch<NearBlocksTokenResponseDto>(`/account/${address}/tokens`);
+        return await this.fetch<string[]>(`/kitwallet/publicKey/${address}/accounts`);
     }
 
     async getLikelyTokens({ address }: NearApiServiceParams): Promise<string[]> {
-        return (await this.getAccountTokens({ address })).tokens.fts;
+        return await this.fetch<string[]>(`/kitwallet/account/${address}/likelyTokens`);
     }
 
     async getLikelyNfts({ address }: NearApiServiceParams): Promise<string[]> {
-        return (await this.getAccountTokens({ address })).tokens.nfts;
+        return await this.fetch<string[]>(`/kitwallet/account/${address}/likelyNFTs`);
     }
 
     async getRecentActivity({ address }: NearApiServiceParams): Promise<Action[]> {
+        try {
+            const activities = await this.fetch<NearBlocksActivityDto[]>(`/kitwallet/account/${address}/activities`);
+
+            return activities.map((activity) => {
+                const transaction: TransactionWithoutActions = {
+                    transactionHash: activity.hash,
+                    includedInBlockHash: activity.block_hash,
+                    blockTimestamp: String(BigInt(activity.block_timestamp) / BigInt(1000000)),
+                    signerAccountId: activity.signer_id,
+                    receiverAccountId: activity.receiver_id,
+                };
+                const action: Action = {
+                    actionKind: this.parseActionKindApiDto(activity.action_kind, activity.receiver_id, address),
+                    transaction: transaction,
+                    transactionHash: activity.hash,
+                    indexInTransaction: activity.action_index,
+                    ...(activity.args.code_sha_256 && { gas: Number(activity.args.code_sha_256) }), // For FUNCTION_CALL kind
+                    ...(activity.args.gas && { gas: Number(activity.args.gas ?? 0) }), // For FUNCTION_CALL kind
+                    ...(activity.args.deposit && { deposit: convertYoctoToNear(activity.args.deposit) }), // For
+                    ...(activity.args.args_base64 && { argsBase64: activity.args.args_base64 }), // For FUNCTION_CALL kind
+                    ...(activity.args.args_json && { argsJson: activity.args.args_json }), // For FUNCTION_CALL kind
+                    ...(activity.args.method_name && { methodName: activity.args.method_name }), // For FUNCTION_CALL kind
+                    ...(activity.args.stake && { stake: convertYoctoToNear(activity.args.stake) }), // For STAKE kind
+                    ...(activity.args.public_key && { publicKey: activity.args.public_key }), // For STAKE, ADD_KEY, DELETE_KEY kind
+                    // ...(activity.args.access_key && { accessKey: activity.args.access_key }), // For ADD_KEY kind
+                    ...(activity.args.beneficiary_id && { beneficiaryId: activity.args.beneficiary_id }), // For DELETE_ACCOUNT kind
+                };
+                return action;
+            });
+        } catch (e) {
+            return [];
+        }
+    }
+
+    async getPaginatedActions({ address, page = 1, pageSize = 10 }: NearApiServicePaginatedParams): Promise<Action[]> {
         const actions = [];
-        let page = 1;
-        const pageSize = 10;
+
         let txns: NearBlocksTransactionDto[] = [];
         do {
             txns = (