+++ src/module/transaction/query/useGetActions.ts
@@ -2,6 +2,7 @@ import { useQuery } from "react-query";
 import useServiceInstance from "module/wallet/hook/useServiceInstance";
 import Queries from "../../../query/queries";
 import { config } from "config";
+import useSelectedWalletIndex from "module/wallet/hook/useSelectedWalletIndex";
 
 export interface UseGetActionsOptions {
     /**
@@ -11,14 +12,16 @@ export interface UseGetActionsOptions {
 }
 
 const useGetActions = ({ index }: UseGetActionsOptions = {}) => {
+    const selectedWallet = useSelectedWalletIndex();
     const { serviceInstance, index: usedIndex, network, queryEnabled } = useServiceInstance(index);
+
     return useQuery(
         [Queries.ACTIONS, usedIndex, network],
         async () => {
             return await serviceInstance.getRecentActivity();
         },
         {
-            enabled: queryEnabled,
+            enabled: queryEnabled && selectedWallet === usedIndex,
             refetchInterval: config.refetchIntervals.transactions,
         },
     );